CL-30 - Code and libraries version refactoring
