+++ app/build.gradle
@@ -70,7 +70,7 @@ android {
 }
 
 dependencies {
-    def material_version = "1.2.0-alpha06"
+    def material_version = "1.3.0-alpha01"
     def constraintlayout_version = "1.1.3"
     def room_version = "2.2.5"
     def preference_version = "1.1.1"
