+++ app/src/main/kotlin/by/liauko/siarhei/cl/recyclerview/holder/RecyclerViewImportFileViewHolder.kt
@@ -1,10 +1,11 @@
-package by.liauko.siarhei.cl.recyclerview
+package by.liauko.siarhei.cl.recyclerview.holder
 
 import android.view.View
 import android.widget.ImageButton
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import by.liauko.siarhei.cl.R
+import by.liauko.siarhei.cl.recyclerview.adapter.RecyclerViewImportFileAdapter
 
 class RecyclerViewImportFileViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
