+++ app/src/main/kotlin/by/liauko/siarhei/cl/util/AppResultCodes.kt
@@ -1,11 +1,10 @@
 package by.liauko.siarhei.cl.util
 
 object AppResultCodes {
-    //TODO: refactor code names
-    const val ADD_LOG = 0x000
-    const val EDIT_LOG = 0x001
-    const val ADD_FUEL_CONSUMPTION = 0x100
-    const val EDIT_FUEL_CONSUMPTION = 0x101
+    const val LOG_ADD = 0x000
+    const val LOG_EDIT = 0x001
+    const val FUEL_CONSUMPTION_ADD = 0x100
+    const val FUEL_CONSUMPTION_EDIT = 0x101
     const val PERIOD_DIALOG_RESULT = 0x200
     const val GOOGLE_SIGN_IN = 0x300
     const val INTERNET_PERMISSION = 0x301
